% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_exposures_threshold}
\alias{generate_exposures_threshold}
\title{Compute exposure values in the four level thresholds model}
\usage{
generate_exposures_threshold(A, Z, threshold, deg = NULL)
}
\arguments{
\item{A}{adj. matrix of a network}

\item{Z}{treatment vector}

\item{threshold}{vector of unit-level thresholds}
}
\value{
vector of observed exposures
}
\description{
Compute exposure values in the four level thresholds model
}
