% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_po}
\alias{generate_po}
\title{Generate PO as in the paper}
\usage{
generate_po(exposures, base.noise = NULL)
}
\arguments{
\item{exposures}{vector of observed exposures}

\item{base.noise}{baseline potential outcome}
}
\description{
Generate PO as in the paper
}
