% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensitivityAnalysis.R
\name{SensitivityAnalysis}
\alias{SensitivityAnalysis}
\title{Sensitivity analysis (SA) for misspecification of the network interference structure}
\usage{
SensitivityAnalysis(
  Z.obs,
  Y.obs,
  theta_vec,
  N_clusters,
  N_each_cluster_vec,
  n.iter,
  distance_matrix = NULL,
  distance.to.prob.func = exp_spatial_distance,
  Pz_function,
  pz_func_args,
  R = 10^4
)
}
\arguments{
\item{Z.obs}{Numeric vector of observed treatments}

\item{Y.obs}{Numeric vector of observed outcomes}

\item{theta_vec}{Vector of Theta parameters to run SA on}

\item{N_clusters}{Number of clusters present}

\item{N_each_cluster_vec}{Numeric vector with size N_clusters that indicate the number of units within each cluster}

\item{n.iter}{Number of SA iteration per Theta value}

\item{distance_matrix}{Optional between-clusters distance matrix; Deafult is NULL}

\item{distance.to.prob.func}{Function that convert distances and Theta value to probailities}

\item{Pz_function}{Function that generate treatment vectors (the experimental design function)}

\item{pz_func_args}{Parameters (as list) required by Pz_function}

\item{R}{Number of re-sampling from Pz_function to estimate the exposures probabilities}
}
\value{
Distribution of the causal effects as a function of theta (as data table)
}
\description{
Return the causal estimates from the obtained by perturbing the specified network
}
\details{
Examine how sensitive causal estimates are to deviations of the specified network;
Suspected misspecification is represented with an induce distribution on the network
Currently only support cross-clusters contamination in cluster-randomized-trials (CRT)
}
\examples{
In the following example, the effects  tau(c11,c00), tau(c10,c00) are estimated
Z.obs <- Z_ber_clusters(N_clusters = 500, N_each_cluster_vec = rep(10,500), p = 0.5)
thres <- runif(5000)
base.Q <- generate_Q_matrix(N_clusters = 500,
                            within_prob_vec = rep(1,500),
                            between_prob_vec = 0.05)
a.mat <- as_adjacency_matrix(sample_sbm(n = 5000, pref.matrix = base.Q, block.sizes = rep(10,500), directed = FALSE))
expos <- generate_exposures_threshold(A = a.mat, Z = Z.obs, threshold = thres)
Y.obs <- generate_po(exposures = expos)
SensitivityAnalysis(Z.obs = Z.obs, Y.obs = Y.obs,
                    theta_vec = seq(0,0.1,0.05),
                    N_clusters = 500,
                    N_each_cluster_vec = rep(10,500),
                    n.iter = 100,
                    Pz_function = Z_ber_clusters,
                    pz_func_args = list(N_clusters = 500,
                                        N_each_cluster_vec = rep(10,500),
                                        p = 0.5),
                    R = 10^4)


}
